HW1/1-2/requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.6.3
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
efficientnet-pytorch==0.7.0
entrypoints==0.3
filelock==3.0.12
flake8==3.8.3
future==0.18.2
idna==2.8
importlib-metadata==2.0.0
importlib-resources==2.0.1
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
olefile==0.46
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1594307298074/work
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
requests==2.21.0
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
stevedore==2.0.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.50.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.24.1
virtualenv==20.0.25
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.2.0
